requirements.txt
flask
python-dotenv
openai
requests
mcp
gunicorn
uvicorn
